pyproject.toml
[project]
name = "dead-by-dawn-sim"
version = "0.1.0"
description = "Combat-focused simulation lab for Dead By Dawn."
readme = "notes/agent_generated/design/simulator_design.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.11.0,<3.0.0",
  "pyyaml>=6.0.2,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "basedpyright>=1.29.1,<2.0.0",
  "prek>=0.3.2,<0.4.0",
  "pytest>=8.3.0,<9.0.0",
  "ruff>=0.11.0,<0.12.0",
  "vulture>=2.14,<3.0",
]

[project.scripts]
ddb-sim = "dead_by_dawn_sim.cli:main"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
cache_dir = ".pytest_cache_local"
addopts = "-p no:cacheprovider"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "C4",
  "C90",
  "TC",
  "UP",
  "N",
  "ARG",
  "PLR",
  "RET",
  "PIE",
  "RUF",
  "SIM",
  "FURB",
]
ignore = [
  "ARG001",
  "E501",
  "PLR0913",
  "PLR2004",
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"src/dead_by_dawn_sim/actions.py" = ["C901", "PLR0911", "PLR0912"]
"src/dead_by_dawn_sim/rules.py" = ["C901", "PLR0912"]
"src/dead_by_dawn_sim/runner.py" = ["C901"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.basedpyright]
include = ["src", "tests"]
pythonVersion = "3.12"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
reportMissingTypeStubs = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
